Refactor game over flow and remove win banner
Replaces the persistent win banner with a timed notification when the game ends. After 10 seconds, the notification is cleared and the lobby state is reset to allow players to ready up for a new game. This change streamlines the end-of-game experience and addresses a user request to remove the win banner.